requirements.txt
pandas>=2.0.0
streamlit>=1.28.0
streamlit-lottie>=0.0.5
pillow>=10.0.0
requests>=2.31.0
email-validator>=2.1.0
python-dotenv>=1.0.0
plotly>=5.17.0


